requirements.txt
Django==3.0.7
psycopg2==2.8.5
djangorestframework==3.11.0
django-cors-headers==3.3.0
django-crispy-forms==1.9.1
django-dynamic-raw-id==2.7
pillow==7.1.2
django-debug-toolbar==2.2
django-extensions==2.2.9
pygraphviz==1.5
django-filter==2.3.0
